Reorganized handlers into new interaction folder
